Changed endpoint URL to make more sense
